test/rgw: refactor tests into multiple scripts
